feat: improve dashboard ui and ux
